requirements.txt
fastapi==0.101.1
email-validator==2.0.0.post2
passlib==1.7.4
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
PyMySQL==1.1.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
uvicorn==0.23.2

